requirements.txt
streamlit-authenticator==0.4.2
matplotlib
seaborn
pandas